+++ packages/unmock-core/src/nock.ts
@@ -1,3 +1,5 @@
+import { array } from "fp-ts/lib/Array";
+import { none, Option, some } from "fp-ts/lib/Option";
 import * as io from "io-ts";
 import { cnst_, extendT, tuple_, type_ } from "json-schema-poet";
 import {
@@ -15,9 +17,11 @@ import {
   JSSTOneOf,
   JSSTTuple,
 } from "json-schema-strictly-typed";
+import { fromTraversable, Iso, Prism } from "monocle-ts";
 import { valAsConst } from "openapi-refinements";
 import * as querystring from "query-string";
 import NodeBackend from "./backend";
+import { identityGetter } from "./generator";
 import { CodeAsInt, HTTPMethod } from "./interfaces";
 import { Schema, ValidEndpointType } from "./service/interfaces";
 import { ServiceStore } from "./service/serviceStore";
@@ -44,6 +48,33 @@ const DynamicJSONValue: io.Type<
   isDynamic,
   (input, context) =>
     isDynamic(input) ? io.success(input) : io.failure(input, context),
+  io.identity,
+);
+
+/*************************************************
+ * (Extended, Maybe) JSON Schema defined below *
+ *************************************************
+ */
+// Used to signal that a value is optional in an object
+const MaybeJSONSymbol: unique symbol = Symbol();
+export interface IMaybeJSONValue {
+  maybe: typeof MaybeJSONSymbol;
+  val: ExtendedValueType;
+}
+const isMaybe = (unk: unknown): unk is IMaybeJSONValue =>
+  typeof unk === "object" &&
+  unk !== null &&
+  (unk as any).maybe === MaybeJSONSymbol &&
+  ExtendedValue.is((unk as any).val);
+
+const MaybeJSONValue: io.Type<IMaybeJSONValue, IMaybeJSONValue> = new io.Type<
+  IMaybeJSONValue,
+  IMaybeJSONValue
+>(
+  "MaybeJSONValueType",
+  isMaybe,
+  (input, context) =>
+    isMaybe(input) ? io.success(input) : io.failure(input, context),
   io.identity,
 );
 
@@ -83,8 +114,59 @@ export type ExtendedValueType =
   | IExtendedObjectType
   | JSONArray
   | JSONObject;
-export interface IExtendedObjectType
+
+export interface ITameExtendedObjectType
   extends Record<string, ExtendedValueType> {} // Defined as interface due to circular reasons
+
+export interface IExtendedObjectType
+  extends Record<string, ExtendedValueType | IMaybeJSONValue> {} // Defined as interface due to circular reasons
+
+const rejectOptionals = (i: IExtendedObjectType): ITameExtendedObjectType =>
+  new Iso<
+    IExtendedObjectType,
+    Array<[string, ExtendedValueType | IMaybeJSONValue]>
+  >(
+    a => Object.entries(a),
+    q => q.reduce((a, b) => ({ ...a, [b[0]]: b[1] }), {}),
+  )
+    .composeTraversal(fromTraversable(array)())
+    .composePrism(
+      new Prism<
+        [string, ExtendedValueType | IMaybeJSONValue],
+        [string, ExtendedValueType]
+      >(
+        a =>
+          ((
+            q: string,
+            r: ExtendedValueType | IMaybeJSONValue,
+          ): Option<[string, ExtendedValueType]> =>
+            MaybeJSONValue.is(r) ? none : some([q, r]))(a[0], a[1]),
+        a => a,
+      ),
+    )
+    .composeGetter(identityGetter())
+    .getAll(i)
+    .reduce((a, b) => ({ ...a, [b[0]]: b[1] }), {});
+
+const keepOptionals = (i: IExtendedObjectType): ITameExtendedObjectType =>
+  new Iso<
+    IExtendedObjectType,
+    Array<[string, ExtendedValueType | IMaybeJSONValue]>
+  >(
+    a => Object.entries(a),
+    q => q.reduce((a, b) => ({ ...a, [b[0]]: b[1] }), {}),
+  )
+    .composeTraversal(fromTraversable(array)())
+    .composePrism(
+      new Prism<
+        [string, ExtendedValueType | IMaybeJSONValue],
+        [string, ExtendedValueType]
+      >(a => (MaybeJSONValue.is(a[1]) ? some([a[0], a[1].val]) : none), a => a),
+    )
+    .composeGetter(identityGetter())
+    .getAll(i)
+    .reduce((a, b) => ({ ...a, [b[0]]: b[1] }), {});
+
 export interface IExtendedArrayType extends Array<ExtendedValueType> {} // Defined as interface due to circular reference
 type EJSEmpty = JSSTEmpty<{}>; // Used as a shortcut
 
@@ -113,81 +195,202 @@ const ExtendedValue: io.Type<
 const ExtendedObject: io.Type<
   IExtendedObjectType,
   IExtendedObjectType
-> = io.recursion("ExtendedObject", () => io.record(io.string, ExtendedValue));
+> = io.recursion("ExtendedObject", () =>
+  io.record(io.string, io.union([ExtendedValue, MaybeJSONValue])),
+);
 const ExtendedArray: io.Type<
   IExtendedArrayType,
   IExtendedArrayType
 > = io.recursion("ExtendedArray", () => io.array(ExtendedValue));
 
-const spreadAndSchemify = (
-  e?: IExtendedObjectType,
-  f: (e: ExtendedValueType) => JSSTAnything<EJSEmpty, {}> = JSONSchemify,
+const spreadAndSchemify = <T, C extends object>(
+  e: ITameExtendedObjectType | undefined,
+  f: (e: ExtendedValueType) => JSSTAnything<T, C>,
 ) =>
   e ? Object.entries(e).reduce((a, b) => ({ ...a, [b[0]]: f(b[1]) }), {}) : {};
 
 // hack until we get around to doing full typing :-(
 const removeDynamicSymbol = (schema: any): JSONSchemaObject<EJSEmpty, {}> => {
-  if (schema instanceof Array) {
-    return (schema as unknown) as JSONSchemaObject<EJSEmpty, {}>;
-  }
-  if (typeof schema === "object") {
-    const { dynamic, ...rest } = schema;
-    return spreadAndSchemify(
-      dynamic === DynamicJSONSymbol ? rest : schema,
-      removeDynamicSymbol,
-    ) as JSONSchemaObject<EJSEmpty, {}>;
-  }
-  return schema;
+  const { dynamic, ...rest } = schema;
+  return rest;
 };
 
-export const JSONSchemify = (
-  e: ExtendedValueType,
+const fuzzNoop = (
+  schema: any,
+): JSONSchemaObject<RecursiveUnionType, IDynamicJSONValue> =>
+  (schema as unknown) as JSONSchemaObject<
+    RecursiveUnionType,
+    IDynamicJSONValue
+  >;
+
+type CType = string | number | boolean | null;
+
+type ConstTransformer<T, C extends object> = (e: CType) => JSSTAnything<T, C>;
+
+// total hack comes from the conversion from schema to json-schema
+// this works because valAsConst only ever yields valid JSON schema
+// we should mitigate this by makeing a "subset" type
+// common to both OAS & JSON Schema
+const simpleConstantTransformer: ConstTransformer<EJSEmpty, {}> = (
+  e: CType,
 ): JSSTAnything<EJSEmpty, {}> =>
+  valAsConst(cnst_<{}>({})(e).const) as JSSTAnything<EJSEmpty, {}>;
+
+const fuzz: ConstTransformer<RecursiveUnionType, IDynamicJSONValue> = (
+  e: CType,
+): JSSTAnything<RecursiveUnionType, IDynamicJSONValue> =>
+  typeof e === "string"
+    ? jspt.string()
+    : typeof e === "number" && Math.floor(e) === e
+    ? jspt.integer()
+    : typeof e === "number"
+    ? jspt.number()
+    : typeof e === "boolean"
+    ? jspt.boolean()
+    : jspt.nul();
+
+const unfuzz: ConstTransformer<RecursiveUnionType, IDynamicJSONValue> = (
+  e: CType,
+): JSSTAnything<RecursiveUnionType, IDynamicJSONValue> =>
+  typeof e === "string"
+    ? jspt.cnst(e)
+    : typeof e === "number"
+    ? jspt.cnst(e)
+    : typeof e === "boolean"
+    ? jspt.cnst(e)
+    : jspt.cnst(null);
+
+export const JSONSchemify = <T, C extends object>(c: C) => (
+  schemaToSchemaTransformer: (schema: any) => JSSTAnything<T, C>,
+) => (constantHandler: ConstTransformer<T, C>) => (
+  e: ExtendedValueType,
+): JSSTAnything<T, C> =>
   isDynamic(e)
-    ? removeDynamicSymbol(
+    ? schemaToSchemaTransformer(
         // we cover all of the nested cases,
         // followed by un-nested cases
         JSSTAllOf(RecursiveUnion, DynamicJSONValue).is(e)
-          ? { ...e, allOf: e.allOf.map(JSONSchemify) }
+          ? {
+              ...e,
+              allOf: e.allOf.map(
+                JSONSchemify<T, C>(c)(schemaToSchemaTransformer)(
+                  constantHandler,
+                ),
+              ),
+            }
           : JSSTAnyOf(RecursiveUnion, DynamicJSONValue).is(e)
-          ? { ...e, anyOf: e.anyOf.map(JSONSchemify) }
+          ? {
+              ...e,
+              anyOf: e.anyOf.map(
+                JSONSchemify<T, C>(c)(schemaToSchemaTransformer)(
+                  constantHandler,
+                ),
+              ),
+            }
           : JSSTOneOf(RecursiveUnion, DynamicJSONValue).is(e)
-          ? { ...e, oneOf: e.oneOf.map(JSONSchemify) }
+          ? {
+              ...e,
+              oneOf: e.oneOf.map(
+                JSONSchemify<T, C>(c)(schemaToSchemaTransformer)(
+                  constantHandler,
+                ),
+              ),
+            }
           : JSSTNot(RecursiveUnion, DynamicJSONValue).is(e)
-          ? { ...e, not: JSONSchemify(e.not) }
+          ? {
+              ...e,
+              not: JSONSchemify<T, C>(c)(schemaToSchemaTransformer)(
+                constantHandler,
+              )(e.not),
+            }
           : JSSTList(RecursiveUnion, DynamicJSONValue).is(e)
-          ? { ...e, items: JSONSchemify(e.items) }
+          ? {
+              ...e,
+              items: JSONSchemify<T, C>(c)(schemaToSchemaTransformer)(
+                constantHandler,
+              )(e.items),
+            }
           : JSSTTuple(RecursiveUnion, DynamicJSONValue).is(e)
-          ? { ...e, oneOf: e.items.map(JSONSchemify) }
+          ? {
+              ...e,
+              oneOf: e.items.map(
+                JSONSchemify<T, C>(c)(schemaToSchemaTransformer)(
+                  constantHandler,
+                ),
+              ),
+            }
           : JSSTObject(RecursiveUnion, DynamicJSONValue).is(e)
           ? {
               ...e,
               ...(e.additionalProperties
-                ? { additionalProperties: JSONSchemify(e.additionalProperties) }
+                ? {
+                    additionalProperties: JSONSchemify<T, C>(c)(
+                      schemaToSchemaTransformer,
+                    )(constantHandler)(e.additionalProperties),
+                  }
                 : {}),
-              ...spreadAndSchemify(e.patternProperties),
-              ...spreadAndSchemify(e.properties),
+              ...(e.patternProperties
+                ? {
+                    patternProperties: spreadAndSchemify(
+                      e.patternProperties,
+                      JSONSchemify<T, C>(c)(schemaToSchemaTransformer)(
+                        constantHandler,
+                      ),
+                    ),
+                  }
+                : {}),
+              ...(e.properties
+                ? {
+                    properties: spreadAndSchemify(
+                      e.properties,
+                      JSONSchemify<T, C>(c)(schemaToSchemaTransformer)(
+                        constantHandler,
+                      ),
+                    ),
+                  }
+                : {}),
             }
           : e,
       )
     : ExtendedArray.is(e) || JSONArray.is(e)
-    ? tuple_<EJSEmpty, {}>({})(e.map(JSONSchemify))
+    ? tuple_<T, C>(c)(
+        e.map(
+          JSONSchemify<T, C>(c)(schemaToSchemaTransformer)(constantHandler),
+        ),
+      )
     : ExtendedObject.is(e) || JSONObject.is(e)
-    ? type_<EJSEmpty, {}>({})(spreadAndSchemify(e), {})
+    ? type_<T, C>(c)(
+        spreadAndSchemify(
+          rejectOptionals(e),
+          JSONSchemify<T, C>(c)(schemaToSchemaTransformer)(constantHandler),
+        ),
+        spreadAndSchemify(
+          keepOptionals(e),
+          JSONSchemify<T, C>(c)(schemaToSchemaTransformer)(constantHandler),
+        ),
+      )
     : e instanceof RegExp
-    ? { type: "string", pattern: e.source }
-    : // total hack comes from the conversion from schema to json-schema
-      // this works because valAsConst only ever yields valid JSON schema
-      // we should mitigate this by makeing a "subset" type
-      // common to both OAS & JSON Schema
-      (valAsConst(cnst_<{}>({})(e).const) as JSSTAnything<EJSEmpty, {}>);
+    ? { type: "string", pattern: e.source, ...c }
+    : constantHandler(e as CType);
 
 // Define poet to recognize the new "dynamic type"
-const jspt = extendT<ExtendedJSONSchema, IDynamicJSONValue>({
+const jspt = extendT<RecursiveUnionType, IDynamicJSONValue>({
   dynamic: DynamicJSONSymbol,
 });
 
-export const u = jspt;
+export const u = {
+  ...jspt,
+  fuzz: JSONSchemify<RecursiveUnionType, IDynamicJSONValue>({
+    dynamic: DynamicJSONSymbol,
+  })(fuzzNoop)(fuzz),
+  unfuzz: JSONSchemify<RecursiveUnionType, IDynamicJSONValue>({
+    dynamic: DynamicJSONSymbol,
+  })(fuzzNoop)(unfuzz),
+  opt: (e: ExtendedJSONSchema): IMaybeJSONValue => ({
+    maybe: MaybeJSONSymbol,
+    val: e,
+  }),
+};
 
 /*******************************
  * Nock-like API defined below *
@@ -247,6 +450,10 @@ interface IDynamicServiceSpec {
   ): IFluentDynamicService & IDynamicServiceSpec;
 }
 
+export const vanillaJSONSchemify = JSONSchemify<EJSEmpty, {}>({})(
+  removeDynamicSymbol,
+)(simpleConstantTransformer);
+
 export class DynamicServiceSpec implements IDynamicServiceSpec {
   private data: Schema = {};
   private headers: Record<string, Schema> = {};
@@ -271,7 +478,10 @@ export class DynamicServiceSpec implements IDynamicServiceSpec {
       ...this.accumulatedQueries,
       ...(data
         ? Object.entries(data).reduce(
-            (a, b) => ({ ...a, [b[0]]: JSONSchemify(b[1]) }),
+            (a, b) => ({
+              ...a,
+              [b[0]]: vanillaJSONSchemify(b[1]),
+            }),
             {},
           )
         : {}),
@@ -306,7 +516,7 @@ export class DynamicServiceSpec implements IDynamicServiceSpec {
     maybeHeaders?: Record<string, InputToPoet>,
   ): IFluentDynamicService & IDynamicServiceSpec {
     if (maybeData !== undefined) {
-      this.data = JSONSchemify(maybeData) as Schema; // TODO should this be some JSSTX?
+      this.data = vanillaJSONSchemify(maybeData) as Schema; // TODO should this be some JSSTX?
       this.statusCode = maybeStatusCode as CodeAsInt | "default";
     } else if (
       (typeof maybeStatusCode === "number" &&
@@ -317,11 +527,14 @@ export class DynamicServiceSpec implements IDynamicServiceSpec {
       // we assume it's a status code
       this.statusCode = maybeStatusCode as CodeAsInt | "default";
     } else {
-      this.data = JSONSchemify(maybeStatusCode) as Schema;
+      this.data = vanillaJSONSchemify(maybeStatusCode) as Schema;
     }
     this.headers = maybeHeaders
       ? (Object.entries(maybeHeaders).reduce(
-          (a, b) => ({ ...a, [b[0]]: JSONSchemify(b[1]) }),
+          (a, b) => ({
+            ...a,
+            [b[0]]: vanillaJSONSchemify(b[1]),
+          }),
           {},
         ) as Record<string, Schema>)
       : {};
@@ -398,7 +611,7 @@ const endpointToQs = (endpoint: ValidEndpointType) =>
   ).reduce(
     (a, b) => ({
       ...a,
-      [b[0]]: JSONSchemify(b[1] === undefined ? null : b[1]),
+      [b[0]]: vanillaJSONSchemify(b[1] === undefined ? null : b[1]),
     }),
     {},
   ) || {};
@@ -459,7 +672,7 @@ const buildFluentNock = (
                   endpointToQs(endpoint) || {},
                   requestHeaders as Record<string, Schema>,
                   requestBody !== undefined
-                    ? (JSONSchemify(requestBody) as Schema)
+                    ? (vanillaJSONSchemify(requestBody) as Schema)
                     : undefined,
                   name,
                 ),