+++ packages/unmock-jest/src/reporter/components/test-suites.tsx
@@ -13,9 +13,10 @@ export const testFileToId = (file: string) => file.replace(/(\/|\\(\\)?|:|\.)/g,
 const TestResults = ({ testSuites }: { testSuites: ITestSuite[] }): [string, () => React.ReactElement] => {
 
     const elementsAndCss = testSuites.map((testSuite, index) => {
-        const id = testFileToId(testSuite.testFilePath);
+        const id = testFileToId(testSuite.shortFilePath);
         const elementCss = `#box-${id}:checked~.test-suite-label-${id} {
     opacity: 1;
+    border-style: solid;
     }
 
     /* Display test suite when corresponding box is checked */
@@ -31,7 +32,7 @@ const TestResults = ({ testSuites }: { testSuites: ITestSuite[] }): [string, () 
         const testSuiteLabelColor = testSuite.suiteResults.numFailingTests > 0 ? `test-suite-label--failure` : `test-suite-label--success`;
         return {
             input: <input type="radio" id={`box-${id}`} className="test-suite-input" name="content" defaultChecked={index === 0} value={`box-${id}`} key={`input-${id}`} />,
-            label: <label htmlFor={`box-${id}`} className={`test-suite-label test-suite-label-${id} ${testSuiteLabelColor}`} id="feature-label" key={`label-${id}`}>
+            label: <label htmlFor={`box-${id}`} className={`test-suite-label test-suite-label-${id} ${testSuiteLabelColor} btn`} id="feature-label" key={`label-${id}`}>
                 <span className="test-suite-label__filename">{testSuite.shortFilePath}</span>
                 <span className="test-suite-label__summary">{`Passing: ${numPassingTests}, failing: ${numFailingTests}, HTTP calls: ${nSnapshots}`}</span>
             </label>,